requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
Flask==2.2.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
prisma==0.9.1 